RunModel/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pandas==2.2.2
scikit-learn==1.5.0
joblib==1.4.2
numpy==1.26.4
lightgbm
nats-py